+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/crud/ProductCrudService.java
@@ -1,8 +1,7 @@
-package com.gulbi.Backend.domain.rental.product.service.product;
+package com.gulbi.Backend.domain.rental.product.service.product.crud;
 
 import com.gulbi.Backend.domain.rental.product.dto.product.ProductDto;
 import com.gulbi.Backend.domain.rental.product.dto.product.ProductOverViewResponse;
-import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductSearchRequestDto;
 import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductUpdateRequestDto;
 import com.gulbi.Backend.domain.rental.product.entity.Product;
