+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/search/strategy/search/ProductSearchStrategy.java
@@ -1,4 +1,4 @@
-package com.gulbi.Backend.domain.rental.product.service.product.strategy.search;
+package com.gulbi.Backend.domain.rental.product.service.product.search.strategy.search;
 
 import com.gulbi.Backend.domain.rental.product.dto.product.ProductOverViewResponse;
