+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/update/ProductUpdatingService.java
@@ -1,10 +1,9 @@
-package com.gulbi.Backend.domain.rental.product.service.product;
+package com.gulbi.Backend.domain.rental.product.service.product.update;
 
 import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductCategoryUpdateRequestDto;
 import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductImageCreateRequestDto;
 import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductImageDeleteRequestDto;
 import com.gulbi.Backend.domain.rental.product.dto.product.request.ProductUpdateRequestDto;
-import com.gulbi.Backend.domain.rental.product.vo.image.ProductImageCollection;
 
 public interface ProductUpdatingService {
     public void updateProductViews(Long productId);
